Add global storage for attributes
